// src/enum/system-message.enum.ts
export enum SystemMessageEnum {
    FILE_NOT_EXIST = 'O arquivo nao existe no local informado \n Por favor, informe no local correto!',
    DATA_VALIDATION_ERROR = 'Os dados estao incorretos!',
    VALID_DATA_INSERT = 'Os dados foram salvos com sucesso!',
    COULD_NOT_DELETE = 'Nao foi possivel remover este cadastro!',
    DATA_REMOVED_SUCCESS = 'Os dados foram removidos com sucesso!',
    DATA_UPDATED_SUCCESS = 'Os dados foram atualizados com sucesso!',
}